Refactor SessionHandler class to improve session handling methods.
