+++ src/web/Session.php
@@ -112,18 +112,85 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
             Yii::warning('Session is already started', __METHOD__);
             $this->updateFlashCounters();
         }
-
-        $this->registerSessionHandler();
     }
 
     public function open(string $path = '', string $name = ''): bool
     {
-        return $this->handler->open($path, $name);
+        if ($this->getIsActive()) {
+            return true;
+        }
+
+        $this->registerSessionHandler();
+        $this->setCookieParamsInternal();
+
+        YII_DEBUG ? session_start() : @session_start();
+
+        if ($this->getUseStrictMode() && $this->_forceRegenerateId) {
+            $this->regenerateID();
+            $this->_forceRegenerateId = null;
+        }
+
+        if ($this->getIsActive()) {
+            Yii::info('Session started', __METHOD__);
+            $this->updateFlashCounters();
+        } else {
+            $error = error_get_last();
+            $message = isset($error['message']) ? $error['message'] : 'Failed to start session.';
+            Yii::error($message, __METHOD__);
+        }
+
+        return true;
     }
 
     public function close(): bool
     {
-        return $this->handler->close();
+        if ($this->getIsActive()) {
+            YII_DEBUG ? session_write_close() : @session_write_close();
+        }
+
+        $this->_forceRegenerateId = null;
+
+        return true;
+    }
+
+    public function destroy(string $id = ''): bool
+    {
+        if ($this->getIsActive()) {
+            $sessionId = session_id();
+
+            $this->close();
+
+            $this->setId($sessionId);
+
+            $this->open('', '');
+
+            session_unset();
+            session_destroy();
+
+            $this->setId($sessionId);
+        }
+
+        return true;
+    }
+
+    /**
+     * @throws InvalidArgumentException if the `$maxLifetime` is invalid.
+     */
+    public function gc(int $maxLifetime): false|int
+    {
+        $this->freeze();
+
+        if ($maxLifetime >= 0 && $maxLifetime <= 100) {
+            // percent * 21474837 / 2147483647 ≈ percent * 0.01
+            ini_set('session.gc_probability', floor($maxLifetime * 21474836.47));
+            ini_set('session.gc_divisor', 2147483647);
+        } else {
+            throw new InvalidArgumentException('GCProbability must be a value between 0 and 100.');
+        }
+
+        $this->unfreeze();
+
+        return 0;
     }
 
     /**
@@ -133,12 +200,12 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
      */
     protected function registerSessionHandler(): void
     {
-        if (is_array($this->handler)) {
-            $this->handler = Yii::createObject($this->handler);
+        if ($this->handler === null) {
+            return;
         }
 
-        if ($this->handler === null) {
-            $this->handler = new SessionHandler($this);
+        if (is_array($this->handler)) {
+            $this->handler = Yii::createObject($this->handler);
         }
 
         if (!$this->handler instanceof SessionHandlerInterface) {
@@ -930,5 +997,35 @@ class Session extends Component implements \IteratorAggregate, \ArrayAccess, \Co
     public function getCacheLimiter(): string
     {
         return session_cache_limiter();
+    }
+
+    /**
+     * Sets the session cookie parameters.
+     * This method is called by [[open()]] when it is about to open the session.
+     *
+     * @throws InvalidArgumentException if the parameters are incomplete.
+     *
+     * @see https://www.php.net/manual/en/function.session-set-cookie-params.php
+     */
+    private function setCookieParamsInternal(): void
+    {
+        $data = $this->getCookieParams();
+
+        if (isset($data['lifetime'], $data['path'], $data['domain'], $data['secure'], $data['httponly'])) {
+            if (!empty($data['samesite'])) {
+                $data['path'] .= '; samesite=' . $data['samesite'];
+            }
+            session_set_cookie_params(
+                $data['lifetime'],
+                $data['path'],
+                $data['domain'],
+                $data['secure'],
+                $data['httponly'],
+            );
+        } else {
+            throw new InvalidArgumentException(
+                'Please make sure cookieParams contains these elements: lifetime, path, domain, secure and httponly.'
+            );
+        }
     }
 }
